+++ tests/sql_in_python_example_correct.py
@@ -2,10 +2,15 @@ query = """
     SELECT * FROM languages;
 """
 
-query = """
+query = '''
     SELECT *
     FROM games
     WHERE test=0
+'''
+
+query = """
+    SELECT * 
+    FROM languages;
 """
 
 
@@ -20,23 +25,38 @@ def get_query2():
     return query
 
 
+def get_query3():
+    limit = 6
+    query = f"""
+    SELECT speed FROM world WHERE animal='dolphin' LIMIT {limit}
+    """
+    return query
+
+
+def get_query4():
+    query = """
+    SELECT speed FROM world WHERE animal='dolphin'
+    """
+    return query
+
+
 def return_5():
     query = 'SELECT 5'
     return query
 
 
 def insert():
-    query = """
+    query = '''
     INSERT INTO table_name (column1, column2, column3)
     VALUES (value1, value2, value3);
-    """
+    '''
     return query
 
 
 def join():
-    query = """
+    query = '''
     SELECT Orders.OrderID, Customers.CustomerName, Orders.OrderDate
     FROM Orders
     INNER JOIN Customers ON Orders.CustomerID=Customers.CustomerID;
-    """
+    '''
     return query